add inplace fix
